get rid of compiler warnings

+++ infra/AnalysisEntry.cpp
@@ -32,7 +32,7 @@ bool AnalysisEntry::ApplyCutOnBranches(std::vector<const Branch*>& br, std::vect
   std::vector<size_t> id_vec;
   bch_vec.reserve(br.size());
   id_vec.reserve(br.size());
-  for (int i = 0; i < br.size(); i++) {
+  for (int i = 0; i < static_cast<int>(br.size()); i++) {
     BranchChannel* bchptr = new BranchChannel(br.at(i), ch.at(i));
     if (cuts.at(i) != nullptr) {
       if (!cuts.at(i)->Apply(*bchptr)) {
@@ -71,7 +71,7 @@ double AnalysisEntry::FillVariable(const Variable& var, std::vector<const Branch
   std::vector<size_t> id_vec;
   bch_vec.reserve(br.size());
   id_vec.reserve(br.size());
-  for (int i = 0; i < br.size(); i++) {
+  for (int i = 0; i < static_cast<int>(br.size()); i++) {
     BranchChannel* bchptr = new BranchChannel(br.at(i), id.at(i));
     bch_vec.emplace_back(bchptr);
     id_vec.emplace_back(br.at(i)->GetId());
